Fix bug in mongodb backend for dead letter queue
In the function{ClaimController:create} for mongodb backend, there
is an error when a message is inserted to the dead letter
queue.Because a mongodb collection(messages) is shared by
many queues, we need to delete first and then insert,
otherwise insert_one will cause the following error:
pymongo.errors.DuplicateKeyError: insertDocument :: caused by ::
11000 E11000 duplicate key error index: admin.test.$_id_  dup key:
{ : ObjectId('xxxxxxxxxxxxxxxxxxxxxx') }

Depends-On: I1ee88a8963e2bc80172710da5ab60313952495e4
Depends-On: I8c3642f3883ecc68853735636fc3a748b9c780b6
Closes-Bug: #1707814

+++ zaqar/storage/mongodb/claims.py
@@ -246,9 +246,13 @@ class ClaimController(storage.Claim):
                                     u"for queue %(dlq_name)s", {"dlq_name":
                                                                 dlq_name})
                         return None, iter([])
-                    result = dlq_collection.insert_one(msg)
-                    if result.inserted_id:
-                        collection.delete_one({'_id': _id})
+                    # NOTE(flwang): If dead letter queue and queue are in the
+                    # same partition, the message has been already
+                    # modified.
+                    if collection != dlq_collection:
+                        result = dlq_collection.insert_one(msg)
+                        if result.inserted_id:
+                            collection.delete_one({'_id': _id})
                     LOG.debug(u"Message %(id)s has met the max claim count "
                               u"%(count)d, now it has been moved to dead "
                               u"letter queue %(dlq_name)s.",
